+++ tests/testthat/test.bq.table.R
@@ -33,3 +33,9 @@ test_that("Correct extension is assigned to a given file format", {
   expect <- "csv.gz"
   expect_equal(res, expect)
 })
+
+test_that("Table can be copied", {
+  expect_true(!bqTableExists("test_table_copy"))
+  bqCopyTable("test_table", "test_table_copy")
+  expect_true(bqTableExists("test_table_copy"))
+})